ssh: Leverage Include directive in config merger
`~/.ssh/config` was created by literally merging the contents of each
file in `~/.ssh/config.d` into `~/.ssh/config`, which means that updates
to the fragments require running `p13n refresh`. By changing the merger
to use the `Include` directive to `~/.ssh/config` (see `man
ssh_config`), `~/.ssh/config` only needs to be updated when a new file
is added to the directory and any changes made to the files take affect
immediately.